backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
supabase>=2.3.0
openai>=1.12.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
PyPDF2>=3.0.0
python-docx>=1.1.0
tiktoken>=0.5.0
numpy>=1.26.0
